agrego modelos y modifico restricciones

+++ src/database/migrations/20231212183140-create-order.js
@@ -28,7 +28,7 @@ module.exports = {
         },
       },
       total: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false
       },
       paymentmethod_id: {
